#21 Move v3 models to v3 package
